tun: disable qdisc queueing
commit 0110d6f22f392f976e84ab49da1b42f85b64a3c5
tun: orphan an skb on tx
Fixed a configuration where skbs get queued
at the tun device forever, blocking senders.

However this fix isn't waterproof:
userspace can control whether the interface
is stopped, and if it is, packets
get queued in the qdisc, again potentially forever.

To complete the fix, this patch sets tun qdisc to 0 when it's registered
to disable qdisc queueing. As user controls this variable to set
queueing within tun as well, set it back to 500 after the
registration.

Signed-off-by: Michael S. Tsirkin <mst@redhat.com>